+++ R/rice.R
@@ -1,3 +1,4 @@
+# could draw.dates and calibratable be modified to work with custom-built as well as prebuilt curves? E.g., using cc=4 for a custom-built curve, cc=5 for another one...
 
 # can intcal.data be made to show the SH data? Currently shows the NH data