+++ NEWS.md
@@ -1,7 +1,11 @@
 # rice 2.2.2
+* the functions `younger`, `older` and `p.range` now also accept age distributions (two columns: ages and their probabilities) or vectors of ages (e.g., rnorm(1e6, 2450, 50)).
 * the functions `younger`, `older` and `p.range` now deal better with BCAD=TRUE.
-* in functions which use `BCAD=TRUE` as option, zero now is an additional option (as sometimes one wishes to include 0 BC/AD, sometimes not).
-* corrected a versioning error in the News.md file of rice version 2.2.1 (some of the items listed in 2.2.1 below were wrongly listed under version 2.2.2)
+* in functions which use `BCAD=TRUE` as option, zero is now an additional option (for those who wish to include 0 BC/AD; default FALSE).
+* corrected a versioning error in the News.md file of rice version 2.2.1 (some of the items listed in 2.2.1 below were wrongly listed under version 2.2.2).
+* the functions `push.gammma` and `push.normal` now invisibly return the distribution (ages and probabilities) as well as the hpd ranges. 
+* cleaned up the messaging of the `fromto` function.
+* new function `create.date` to simulate a radiocarbon date from a calendar date. This looks up the calibration curve's C14 age and error, adds any scatter and reservoir effect, and uses a 2 permille laboratory error by default.  
 
 # rice 2.2.1
 * `draw.dates` now plots calibrated dates correctly also if cc>1.